♻️ [generic] Resolves #9 by spliting monitor into a base pollster class and monitor derived class
