+++ include/jage/input/cursor/monitor.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <jage/input/cursor/state.hpp>
-#include <jage/input/generic/monitor.hpp>
+#include <jage/input/generic/pollster.hpp>
 
 #include <cstddef>
 
@@ -13,13 +13,18 @@ static constexpr auto update_state =
   state = driver.cursor_position();
 };
 
+static constexpr auto invoke_callback =
+    [](const auto &callback, const jage::input::cursor::state &state) -> void {
+  callback(state);
+};
+
 template <class TDriver, std::size_t CallbackCapacity = 2>
-using monitor =
-    generic::monitor<TDriver, CallbackCapacity, state, std::uint8_t,
-                     update_state, [](const std::uint8_t &) -> void {},
-                     [](const auto &callback,
-                        const jage::input::cursor::state &state) -> void {
-                       callback(state);
-                     }>;
+struct monitor
+    : public generic::pollster<TDriver, CallbackCapacity, state, update_state,
+                               invoke_callback, state> {
+  explicit monitor(TDriver &driver)
+      : generic::pollster<TDriver, CallbackCapacity, state, update_state,
+                          invoke_callback, state>{driver} {}
+};
 
 } // namespace jage::input::cursor
